- Heuristics.*:   . Corregido update para que actualize correctamente los     bins usados y la cantidad de bins restantes
- utils.*:
  . Corregida combinations para que no se saliera del rango
    y para que no devolviera combinaciones fuera de su dominio.
  . Agregada searchBin, para ver si un bin esta en un packing y donde

- HFirstBest.*:
  . bug en el casteo a long de combs

- LocalSearch.*:
  . Ahora filling devuelve un double, para mayor precision

- binPacking2D.cpp:
  . Ahora se llama a LocalSearch

- Ya se acoplo HFirstBest

+++ LocalSearch.h
@@ -9,6 +9,6 @@ Packing LocalSearch(vector<Item>,int,int);
 
 int targetBin(Packing,int,int);
 
-int filling(Packing,int,int,int);
+double filling(Packing,int,int,int);
 
 Packing initialPacking(vector<Item>);
